Add dynamic types reading

+++ cuda/src/include/types.h
@@ -3,10 +3,10 @@
 
 // Define your precision types
 // Intended to use RealTypeConstant for DT, M, Et, sigmaL 
-using RealTypeConstant = double;
+using RealTypeConstant = float;
 
 // Intended to use RealTypeConstant for d_x, h_x, d_DxE 
-using RealTypeX = double;
+using RealTypeX = float;
 
 // Use RealType2 for d_p, DpE, h_p, 
 using RealTypeP = double;
